Add the delete family command handler

+++ FamilyGallery.Application/Features/Families/Commands/DeleteFamily/DeleteFamilyCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using FamilyGallery.Application.Contracts.Persistence;
+using FamilyGallery.Application.Exceptions;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -9,7 +10,7 @@ using System.Threading.Tasks;
 
 namespace FamilyGallery.Application.Features.Families.Commands.DeleteFamily
 {
-    public class DeleteFamilyCommandHandler : IRequestHandler<DeleteFamilyCommand>
+    public class DeleteFamilyCommandHandler : IRequestHandler<DeleteFamilyCommand, DeleteFamilyCommandResponse>
     {
         private readonly IFamilyRepository familyRepository;
 
@@ -17,11 +18,17 @@ namespace FamilyGallery.Application.Features.Families.Commands.DeleteFamily
         {
             this.familyRepository = familyRepository;
         }
-        public async Task<Unit> Handle(DeleteFamilyCommand request, CancellationToken cancellationToken)
+        public async Task<DeleteFamilyCommandResponse> Handle(DeleteFamilyCommand request, CancellationToken cancellationToken)
         {
-            var family = await familyRepository.GetByIdAsync(request.Id);
+            var validator = new DeleteFamilyCommandValidator(familyRepository);
+            var validationResult = await validator.ValidateAsync(request);
+            if (!validationResult.IsValid)
+            {
+                return new DeleteFamilyCommandResponse { IsSuccessful = true, Message = validationResult.ToString() };
+            }
+            var family = await familyRepository.GetByIdAsync(request.FamilyId);
             await familyRepository.DeleteAsync(family);
-            return Unit.Value;
+            return new DeleteFamilyCommandResponse { IsSuccessful = true };
         }
     }
 }
